split non-field definitions into separate files
