+++ lib/db_schema/definitions/check_constraint.rb
@@ -1,0 +1,13 @@
+module DbSchema
+  module Definitions
+    class CheckConstraint
+      include Dry::Equalizer(:name, :condition)
+      attr_reader :name, :condition
+
+      def initialize(name:, condition:)
+        @name      = name.to_sym
+        @condition = condition
+      end
+    end
+  end
+end
