+++ lib/db_schema/definitions/extension.rb
@@ -1,0 +1,12 @@
+module DbSchema
+  module Definitions
+    class Extension
+      include Dry::Equalizer(:name)
+      attr_reader :name
+
+      def initialize(name)
+        @name = name
+      end
+    end
+  end
+end
